Applied dropdown reset filter functionality.
Upon change of dropdown, the clickData is reset to None. This should fix the index out of bounds error. After clicking the housepie (i.e., click Pangasinan, so we have clickData = "Pangasinan", it is only Pangasinan that is in the pie chart and the graphs tied to it change accordingly). If you click the dropdown of changing the region, clickData = None.